refactor: add useTodoInput hooks
